Remove pxPath, rename [animation] to [pxAnimation], create [pxObject] file
