+++ examples/pxScene2d/src/pxAnimate.cpp
@@ -18,6 +18,7 @@
 
 // pxAnimate.cpp
 
+#include "pxObject.h"
 #include "pxAnimate.h"
 #include "pxScene2d.h"
 
@@ -97,7 +98,7 @@ void pxAnimate::setStatus (pxConstantsAnimation::animationStatus status)
     mStatus = status;
 }
 
-void pxAnimate::update (const char* prop, struct animation* params, pxConstantsAnimation::animationStatus status)
+void pxAnimate::update (const char* prop, struct pxAnimation* params, pxConstantsAnimation::animationStatus status)
 {
   if (NULL != mCurrDetails.getPtr())
   {
